+++ CHANGELOG.md
@@ -111,6 +111,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Deleting 0 columns is no longer an error.
 - Rows filter now works properly when applied to a view table and using "eager"
   evaluation engine.
+- Computed columns expression can now be combined with rows expression, or
+  applied to a view Frame.
 
 
 ### [v0.2.2](https://github.com/h2oai/datatable/compare/v0.2.2...v0.2.1) — 2017-10-18